Refactor AgentCore Lambda deployment to extract reusable components
Extracted webhook routing and app adapter logic from main.py and index.py
into reusable modules for better maintainability and code organization.

Changes:
- Added credentials.py: Twilio credential fetching from Secrets Manager
- Added proxy.py: AgentCoreLambdaProxy for Lambda webhook routing
- Added tac_app.py: TACAgentCoreApp adapter for BedrockAgentCoreApp integration
- Simplified main.py: Removed app adapter boilerplate (100+ lines)
- Simplified index.py: Removed webhook routing logic (190+ lines)

Benefits:
- Improved separation of concerns
- Reusable components for future deployments
- Better testability with isolated modules
- Cleaner main entry points
