+++ deploy/agentcore_aws_lambda/lambda/proxy.py
@@ -1,0 +1,250 @@
+"""
+AgentCore Lambda Proxy - Reusable Lambda handler wrapper for TAC integration.
+
+This module provides a proxy class that wraps Lambda handler logic for Twilio webhook
+routing to AWS Bedrock AgentCore. Users provide the Twilio auth token, and the proxy
+handles signature validation and webhook routing internally.
+
+Example usage:
+    from credentials import fetch_twilio_auth_token
+    from proxy import AgentCoreLambdaProxy
+
+    # Fetch Twilio auth token from Secrets Manager
+    twilio_auth_token = fetch_twilio_auth_token(
+        secret_arn=os.environ["TWILIO_SECRET_ARN"]
+    )
+
+    # Create proxy
+    proxy = AgentCoreLambdaProxy(
+        agentcore_runtime_arn=os.environ["AGENTCORE_RUNTIME_ARN"],
+        conversation_configuration_id=os.environ["TWILIO_CONVERSATION_CONFIGURATION_ID"],
+        twilio_auth_token=twilio_auth_token,
+    )
+
+    # Expose handler
+    lambda_handler = proxy.lambda_handler
+"""
+
+import json
+import os
+from typing import Any
+
+import boto3
+from bedrock_agentcore.runtime import AgentCoreRuntimeClient
+from tac.channels.voice.twiml import generate_twiml
+from tac.core.logging import get_logger
+from validation import TwilioSignatureValidator
+
+logger = get_logger(__name__)
+
+
+class AgentCoreLambdaProxy:
+    """Lambda proxy for routing Twilio webhooks to AWS Bedrock AgentCore.
+
+    This proxy handles:
+    - Signature validation for Twilio webhooks
+    - Voice call routing (generates TwiML with presigned WebSocket URL)
+    - Conversation webhook routing (forwards to AgentCore via HTTP)
+
+    Attributes:
+        agentcore_runtime_arn: ARN of the AgentCore runtime
+        conversation_configuration_id: Twilio Conversation Configuration ID
+        signature_validator: TwilioSignatureValidator instance
+        aws_region: AWS region for clients
+    """
+
+    def __init__(
+        self,
+        agentcore_runtime_arn: str,
+        conversation_configuration_id: str,
+        twilio_auth_token: str,
+        aws_region: str | None = None,
+    ):
+        """Initialize AgentCore Lambda proxy.
+
+        Args:
+            agentcore_runtime_arn: ARN of the AgentCore runtime
+            conversation_configuration_id: Twilio Conversation Configuration ID
+            twilio_auth_token: Twilio auth token for webhook signature validation
+            aws_region: AWS region for boto3 clients (optional, defaults to AWS_REGION
+                environment variable provided by Lambda runtime)
+        """
+        self.agentcore_runtime_arn = agentcore_runtime_arn
+        self.conversation_configuration_id = conversation_configuration_id
+        self.signature_validator = TwilioSignatureValidator(twilio_auth_token)
+        self.aws_region = aws_region or os.environ["AWS_REGION"]
+
+        self.agent_core_client = boto3.client("bedrock-agentcore", region_name=self.aws_region)
+        self.agentcore_runtime_client = AgentCoreRuntimeClient(region=self.aws_region)
+
+    def lambda_handler(self, event: dict[str, Any], context: Any) -> dict[str, Any]:
+        """Route requests to appropriate handler.
+
+        Args:
+            event: Lambda event containing HTTP request data
+            context: Lambda context object
+
+        Returns:
+            Dict containing statusCode, headers, and body
+        """
+        if not self.signature_validator.validate(event):
+            return {
+                "statusCode": 403,
+                "headers": {"Content-Type": "application/json"},
+                "body": json.dumps({"error": "Invalid webhook signature"}),
+            }
+
+        request_path = event.get("rawPath", event.get("path", "/"))
+
+        if request_path.startswith("/twiml"):
+            return self._handle_voice_twiml(event)
+        elif request_path.startswith("/webhook"):
+            return self._handle_conversation_webhook(event)
+        else:
+            return {"statusCode": 404, "body": json.dumps({"error": "Not found"})}
+
+    def _handle_voice_twiml(self, event: dict[str, Any]) -> dict[str, Any]:
+        """Handle voice call - generate TwiML.
+
+        Uses CallSid as session_id for AgentCore routing to ensure the call is routed
+        to a dedicated agent instance based on CallSid.
+
+        Args:
+            event: Lambda event containing Twilio voice webhook data
+
+        Returns:
+            Dict containing statusCode 200 with TwiML body, or error response
+        """
+        try:
+            call_sid = self._extract_call_sid(event)
+            if not call_sid:
+                logger.warning("Missing CallSid in voice webhook request")
+                return {
+                    "statusCode": 400,
+                    "headers": {"Content-Type": "application/json"},
+                    "body": json.dumps({"error": "Missing CallSid parameter"}),
+                }
+
+            websocket_url = self.agentcore_runtime_client.generate_presigned_url(
+                runtime_arn=self.agentcore_runtime_arn, session_id=call_sid, expires=300
+            )
+
+            twiml = generate_twiml(
+                options={
+                    "websocket_url": websocket_url,
+                    "conversation_configuration": self.conversation_configuration_id,
+                }
+            )
+
+            return {
+                "statusCode": 200,
+                "headers": {"Content-Type": "application/xml"},
+                "body": twiml,
+            }
+
+        except Exception as e:
+            logger.error(f"Error generating TwiML: {str(e)}", exc_info=True)
+            return {
+                "statusCode": 500,
+                "headers": {"Content-Type": "application/json"},
+                "body": json.dumps({"error": str(e)}),
+            }
+
+    def _extract_call_sid(self, event: dict[str, Any]) -> str | None:
+        """Extract CallSid from Lambda event (POST body or query string).
+
+        Args:
+            event: Lambda event containing HTTP request data
+
+        Returns:
+            CallSid string if found, None otherwise
+        """
+        body = event.get("body", "")
+        if body:
+            from urllib.parse import parse_qs
+
+            if event.get("isBase64Encoded"):
+                import base64
+
+                body = base64.b64decode(body).decode("utf-8")
+            params = parse_qs(body)
+            call_sid = params.get("CallSid", [None])[0]
+            if call_sid:
+                return call_sid
+
+        query_params = event.get("queryStringParameters") or {}
+        return query_params.get("CallSid")
+
+    def _handle_conversation_webhook(self, event: dict[str, Any]) -> dict[str, Any]:
+        """Handle conversation webhook - forward to AgentCore.
+
+        Conversation Orchestrator sends JSON webhooks. TAC processes only:
+        - COMMUNICATION_CREATED: New message from customer (data.conversationId)
+        - CONVERSATION_UPDATED: Conversation status changed (data.id)
+
+        Other event types are ignored.
+
+        Args:
+            event: Lambda event containing Conversation Orchestrator webhook data
+
+        Returns:
+            Dict containing statusCode and success response
+        """
+        try:
+            body = event.get("body", "")
+            headers = event.get("headers", {})
+
+            webhook_data = json.loads(body)
+            event_type = webhook_data.get("eventType")
+            event_data = webhook_data.get("data", {})
+
+            if event_type not in ("COMMUNICATION_CREATED", "CONVERSATION_UPDATED"):
+                logger.debug(f"Ignoring event type: {event_type}")
+                return {
+                    "statusCode": 200,
+                    "headers": {"Content-Type": "application/json"},
+                    "body": json.dumps({"success": True, "ignored": True}),
+                }
+
+            # Extract conversation_id (location differs by event type)
+            conversation_id = None
+            if event_type == "COMMUNICATION_CREATED":
+                conversation_id = event_data.get("conversationId")
+            elif event_type == "CONVERSATION_UPDATED":
+                conversation_id = event_data.get("id")
+
+            if not conversation_id:
+                logger.warning(f"Missing conversationId for event type {event_type}")
+                return {
+                    "statusCode": 400,
+                    "headers": {"Content-Type": "application/json"},
+                    "body": json.dumps({"error": "Missing conversationId"}),
+                }
+
+            payload = json.dumps(
+                {
+                    "webhook_data": body,
+                    "idempotency_token": headers.get("i-twilio-idempotency-token"),
+                }
+            ).encode()
+
+            # Use conversation_id as runtimeSessionId for sticky routing to same microVM
+            self.agent_core_client.invoke_agent_runtime(
+                agentRuntimeArn=self.agentcore_runtime_arn,
+                runtimeSessionId=conversation_id,
+                payload=payload,
+            )
+
+            return {
+                "statusCode": 200,
+                "headers": {"Content-Type": "application/json"},
+                "body": json.dumps({"success": True}),
+            }
+
+        except Exception as e:
+            logger.error(f"Error forwarding webhook to AgentCore: {str(e)}", exc_info=True)
+            return {
+                "statusCode": 500,
+                "headers": {"Content-Type": "application/json"},
+                "body": json.dumps({"error": str(e)}),
+            }
